Refactor: Create m.parsing package
