+++ Code/m/src/m/validation/MValidator.java
@@ -797,243 +797,4 @@ public class MValidator extends AbstractMValidator
 			}
 		}
 	}
-	
-	/*
-	@Check
-	public void infer(EntityVariable variable)
-	{
-		var component = variable.getComponent();
-		if (component != null)
-		{
-			if (component.endsWith("Range"))
-			{
-				set(component.substring(0, component.indexOf("Range")), input);
-				set(component, float1);
-			}
-			else if (component.endsWith("Vector"))
-			{
-				set(component.substring(0, component.indexOf("Vector")), input);
-				set(component, float2);
-			}
-		}
-	}
-	
-	@Check
-	public void infer(Loop loop)
-	{
-		var entity = loop.getEntity();
-		set(entity, Type.entity);
-		
-		var collection = (EntityVariable) loop.getCollection();
-		if (collection != null)
-		{
-			set(collection.getEntity(), Type.entity);
-			set(collection.getComponent(), Type.entityList);
-		}
-		
-		var tags = loop.getTags();
-		for (var t : tags)
-		{
-			set(t, Type.tag);
-		}
-	}
-	
-	@Check
-	public void infer(Selection selection)
-	{
-		for (var branch : selection.getBranches())
-		{
-			var condition = branch.getCondition();
-			if (condition != null)
-			{
-				set(condition, Type.bool);
-			}
-		}
-	}
-	
-	@Check
-	public void infer(Iteration iteration)
-	{
-		var condition = iteration.getCondition();
-		set(condition, Type.bool);
-	}
-	
-	@Check
-	public void infer(FunctionCall functionCall)
-	{
-		var name = functionCall.getName();
-		var parameters = functionCall.getParameters();
-		
-		var mathFunctions = new ArrayList<String>();
-		mathFunctions.add("sin");
-		mathFunctions.add("cos");
-		mathFunctions.add("tan");
-		mathFunctions.add("exp");
-		mathFunctions.add("log");
-		
-		if (mathFunctions.contains(name))
-		{
-			set(parameters.get(0), float1);				
-		}
-		else if (name.equals("random"))
-		{
-			set(parameters.get(0), float2);
-		}
-		else if (name.equals("create"))
-		{
-			set(parameters.get(0), entity);
-		}
-		else if (name.equals("destroy"))
-		{
-			set(parameters.get(0), entity);
-		}
-		else if (name.equals("add")||name.equals("remove"))
-		{
-			set(parameters.get(1), entity);
-		}
-	}
-	
-	@Check
-	public void infer(Assignment assignment)
-	{
-		var variable = (EntityVariable) assignment.getVariable();
-		var name = variable.getEntity();
-		if (variable.getComponent() != null)
-		{
-			name = variable.getComponent();
-		}
-		var expression = assignment.getExpression();
-		switch(assignment.getKind())
-		{
-			case AND: 
-			{
-				set(name, float1);
-				set(expression,float1);
-				break;
-			}
-			case DECREASE: 
-			{
-				group(name, expression);
-				break;
-			}
-			case DIVIDE: 
-			{
-				set(expression,float1);
-				break;
-			}
-			case INCREASE:
-			{
-				group(name, expression);
-				break;
-			}
-			case MODULUS: 
-			{
-				set(name, float1);
-				set(expression,float1);
-				break;
-			}
-			case MULTIPLY: 
-			{
-				set(expression,float1);
-				break;
-			}
-			case OR: 
-			{
-				set(name, float1);
-				set(expression,float1);
-				break;
-			}
-			case SET: 
-			{
-				group(name, expression);
-				break;
-			}
-			case SHIFT_LEFT: 
-			{
-				set(name, float1);
-				set(expression,float1);
-				break;
-			}
-			case SHIFT_RIGHT: 
-			{
-				set(name, float1);
-				set(expression,float1);
-				break;
-			}
-			case XOR: 
-			{
-				set(name,float1);
-				set(expression,float1);
-				break;
-			}
-		}
-	}
-	
-	private void set(Expression expression, Type type)
-	{
-		
-	}
-	
-	private void group(String name, Expression expression)
-	{
-		
-	}
-	
-	@Check
-	public void check(Addition addition)
-	{
-		var left = (EntityVariable) addition.getLeft();
-		var right = (EntityVariable) addition.getRight();
-		var leftComponent = left.getComponent();
-		var rightComponent = right.getComponent();
-		
-		if (!types.containsKey(leftComponent)||!types.containsKey(rightComponent))
-		{
-			warning("components not typed", ModularPackage.Literals.ADDITION__LEFT);
-		}
-		var l = types.get(leftComponent);
-		var r = types.get(rightComponent);
-		
-		if (!isNumeral(l))
-		{
-			error("left is not numeral", ModularPackage.Literals.ADDITION__LEFT);
-		}
-		
-		if (l != r)
-		{
-			error("types dont match", ModularPackage.Literals.ADDITION__LEFT);
-		}
-	}
-	
-	@Check
-	public void check(Assignment assignment)
-	{
-		var left = (EntityVariable) assignment.getVariable();
-		var name = left.getEntity();
-		if (left.getComponent() != null)
-		{
-			
-		}
-	}
-	
-	private boolean isNumeral(Type t)
-	{
-		return t == float1 || t == float2 || t == Type.float3 || t == Type.float4;
-	}
-	
-	private void set(String name, Type type)
-	{
-		if (types.containsKey(name))
-		{
-			if (types.get(name) != type)
-			{
-				error("Type error, expected " + types.get(name) + " but was " + type,MPackage.Literals.COMPONENT__NAME);
-			}
-		}
-		else
-		{
-			types.put(name, type);
-		}
-	}*/
-	
 }
